[toolbars-new] refactor: clean buttonListCreator
