--- packages/toolbars-new/web/src/Toolbar/utils.ts
@@ -1,15 +1,0 @@
-import { DROPDOWN_OPTIONS_TO_DOC_STYLE_TYPE, DocumentStyle } from 'wix-rich-content-common';
-
-export const getFontSizeNumber = (fontSize: string) => {
-  const pxRegex = new RegExp('[0-9]+[px]');
-  return pxRegex.exec(fontSize) ? fontSize.split('p')[0] : '';
-};
-
-export const hasStyleChanges = (
-  currentHeading: string,
-  inlineStyles: Record<string, string>,
-  documentStyle?: DocumentStyle
-) => {
-  const headerStyle = documentStyle?.[DROPDOWN_OPTIONS_TO_DOC_STYLE_TYPE[currentHeading]] || {};
-  return Object.entries(inlineStyles).some(([key, value]) => headerStyle[key] !== value);
-};
